+++ src/Components/Validator/CaseStyleValidator.php
@@ -4,34 +4,21 @@ namespace PHPUnuhi\Components\Validator;
 
 use PHPUnuhi\Bundles\Storage\StorageInterface;
 use PHPUnuhi\Components\Validator\CaseStyle\CaseStyleValidatorFactory;
+use PHPUnuhi\Components\Validator\Model\ValidationError;
+use PHPUnuhi\Components\Validator\Model\ValidationResult;
 use PHPUnuhi\Models\Translation\TranslationSet;
-use Symfony\Component\Console\Output\OutputInterface;
 
 class CaseStyleValidator implements ValidatorInterface
 {
-
-    /**
-     * @var OutputInterface
-     */
-    private $output;
-
-
-    /**
-     * @param OutputInterface $output
-     */
-    public function __construct(OutputInterface $output)
-    {
-        $this->output = $output;
-    }
 
 
     /**
      * @param TranslationSet $set
      * @param StorageInterface $storage
-     * @return bool
-     * @throws \Exception
+     * @return ValidationResult
+     * @throws CaseStyle\Exception\CaseStyleNotFoundException
      */
-    public function validate(TranslationSet $set, StorageInterface $storage): bool
+    public function validate(TranslationSet $set, StorageInterface $storage): ValidationResult
     {
         $caseValidatorFactory = new CaseStyleValidatorFactory();
 
@@ -40,7 +27,8 @@ class CaseStyleValidator implements ValidatorInterface
         $hierarchy = $storage->getHierarchy();
 
 
-        $isValid = true;
+        $validationErrors = [];
+
 
         foreach ($set->getCasingStyles() as $style) {
             $caseValidators[] = $caseValidatorFactory->fromIdentifier($style);
@@ -85,22 +73,19 @@ class CaseStyleValidator implements ValidatorInterface
                 }
 
                 if (!$isKeyCaseValid) {
-
-                    $this->output->writeln("[CASE-STYLE] Found invalid case-style in locale: " . $locale->getName());
-                    if (!empty($locale->getFilename())) {
-                        $this->output->writeln("  - " . $locale->getFilename());
-                    }
-
-                    $this->output->writeln('           [x]: ' . $translation->getKey());
-                    $this->output->writeln("");
-
-                    $isValid = false;
+                    $validationErrors[] = new ValidationError(
+                        'CASE-STYLE',
+                        'Invalid case-style for key',
+                        $locale->getName(),
+                        $locale->getFilename(),
+                        $translation->getKey()
+                    );
                     break;
                 }
             }
         }
 
-        return $isValid;
+        return new ValidationResult($validationErrors);
     }
 
 }